refactor: rename route.ts to processTranscript.ts
